wikipedia/requirements.txt
feedparser
python-dotenv==0.20.0
requests
chardet
confluent_kafka
